[style] add new file close\/open_pipes.c
