+++ Makefile
@@ -6,7 +6,7 @@
 #    By: yliu <yliu@student.42.jp>                  +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2023/11/09 12:04:47 by yliu              #+#    #+#              #
-#    Updated: 2024/05/05 11:22:08 by yliu             ###   ########.fr        #
+#    Updated: 2024/05/05 13:52:44 by yliu             ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -60,6 +60,7 @@ LIB				:= $(LIB_DIR)/$(LIBRARY)
 
 ORIGIN_HEADERS	:=	./inc/$(NAME).h \
 					./inc/main_helper.h \
+					./inc/error.h \
 					./inc/utils.h \
 					./inc/dlist.h \
 					./inc/xwrapper.h \
@@ -68,7 +69,9 @@ ORIGIN_HEADERS	:=	./inc/$(NAME).h \
 					./inc/parse_string.h
 
 BASIC_SRCS 		:=	./src/main/init_info.c \
-					./src/main/main_helper/pipe_utils.c \
+					./src/utils/pipe_utils.c \
+					./src/main/open_pipes.c \
+					./src/main/close_pipes.c \
 					./src/main/loop_xfork.c \
 					\
 					./src/here_doc/here_doc.c \
