+++ inc/error.h
@@ -1,30 +1,30 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   pipex.h                                            :+:      :+:    :+:   */
+/*   error.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: yliu <yliu@student.42.jp>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/02/05 18:16:51 by yliu              #+#    #+#             */
-/*   Updated: 2024/05/05 13:20:45 by yliu             ###   ########.fr       */
+/*   Created: 2024/05/05 13:46:22 by yliu              #+#    #+#             */
+/*   Updated: 2024/05/05 13:50:45 by yliu             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef PIPEX_H
-# define PIPEX_H
+#ifndef ERROR_H
+# define ERROR_H
 
-# include "dlist.h"
-# include "ft_printf.h"
-# include "get_next_line.h"
-# include "here_doc.h"
-# include "libft.h"
-# include "main_helper.h"
-# include "parse_string.h"
-# include "utils.h"
-# include "xfork_exec.h"
-# include "xwrapper.h"
-# include <fcntl.h>
+# include <sys/errno.h>
 
-int	main(int argc, const char **argv, const char *envp[]);
+# define CMD_NOT_FOUND "command not found"
+
+typedef enum e_exit_status
+{
+	NOT_EXECUTABLE = 126,
+	NO_FILE_OR_CMD_ERR = 127
+}	t_exit_status;
+
+int	print_error(const char *basename, char *str, int exit_status);
+int	print_errno(const char *errno_msg);
+int	print_quote_error(void);
 
 #endif
