+++ inc/main_helper.h
@@ -6,7 +6,7 @@
 /*   By: yliu <yliu@student.42.jp>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/04 14:19:15 by yliu              #+#    #+#             */
-/*   Updated: 2024/05/05 13:14:12 by yliu             ###   ########.fr       */
+/*   Updated: 2024/05/05 13:34:44 by yliu             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,6 +14,7 @@
 # define MAIN_HELPER_H
 
 # include "here_doc.h"
+# include "xfork_exec.h"
 # include "xwrapper.h"
 # include <fcntl.h>
 
@@ -43,9 +44,11 @@ typedef struct s_arg
 	bool		is_heredoc;
 }				t_arg;
 
-void			mk_xpipe(int *pip_arr, int i);
-int				pipe_read_fd(int *pipefd, int i);
-int				pipe_write_fd(int *pipefd, int i);
 void			init_arg_info(int argc, const char **argv, const char **envp,
 					t_arg *arg_info);
+void			open_pipes(int cmd_i, int *pipefd, t_arg *arg_info,
+					t_fd *fd_info);
+void			close_pipes(int cmd_i, int *pipefd, t_arg *arg_info,
+					t_fd *fd_info);
+void			loop_xfork(t_arg *arg_cve_info, t_fd *fd_info, int *pipefd);
 #endif
