+++ src/utils/pipe_utils.c
@@ -6,11 +6,11 @@
 /*   By: yliu <yliu@student.42.jp>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/04 14:18:47 by yliu              #+#    #+#             */
-/*   Updated: 2024/05/05 10:17:08 by yliu             ###   ########.fr       */
+/*   Updated: 2024/05/05 13:35:05 by yliu             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "main_helper.h"
+#include "utils.h"
 
 void	mk_xpipe(int *pip_arr, int i)
 {
